refactor(component): rename Table -> DisplayTable
ref: #133 #132
